backend_api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
requests==2.31.0
python-multipart==0.0.6
jinja2==3.1.3
httpx==0.26.0
beautifulsoup4==4.12.3
pillow==10.2.0
dnspython==2.5.0
phonenumbers==8.13.29
whois==0.9.27
google-generativeai==0.3.2
python-docx==1.1.0
trio==0.24.0
vysion==0.0.5 # Assuming this version or verifying availability
